Split up work for bootstrap precache
As mentioned in #29, submitting all tasks to the executor
implies that the executor will have limited # of threads.

But if used with a cached thread pool, we will easily hit 1k threads+

This manually splits the work based on core count to a reasonable
split rate, and then if the submitter wants to limit threads
even more, it can pass an executor that limits thread count.